feat: show bookings separately
