+++ catcare_front/src/App.styles.ts
@@ -2,11 +2,12 @@ import styled from 'styled-components'
 
 export const AppContainer = styled.div`
   padding: 10px 0;
-  max-height: 100vh;
-  min-height: 100vh;
-  height: 100%;
+  height: 100vh;
   max-width: 430px;
   width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   overflow-y: auto;
 
   &::-webkit-scrollbar {
