+++ catcare_front/src/screens/BookingsList/index.ts
@@ -1,0 +1,1 @@
+export * from './BookingsList'
