+++ catcare_front/src/screens/Home/SitterHomeScreen.tsx
@@ -3,32 +3,17 @@ import { useContext, useEffect, useState, useCallback } from 'react'
 import axios from 'axios'
 import { AuthContext } from '@/context'
 import { toast } from 'react-toastify'
-import { mockedUserBookings, requestersData } from './utils'
-import {
-  BookingCard,
-  CardsList,
-  DateInfo,
-  Footer,
-  Header,
-  Address,
-  Name,
-  Notes,
-  SitterHomeContainer,
-  VisitWrapper,
-  VisitSummary,
-  IconButton,
-  VisitItem,
-  Info,
-  Details
-} from './SitterStyles'
+import { mockedUserBookings } from './utils'
+import { SitterHomeContainer, ButtonsWrapper } from './SitterStyles'
 import { Booking } from '@/domain/models/Booking'
-import { ArrowDropDown } from '@mui/icons-material'
+import { BookingsList } from '../BookingsList'
 
 function SitterHomeScreen() {
   const { getAuthTokenFromStorage, authState } = useContext(AuthContext)
 
   const [pendingBookings, setPendingBookings] = useState<Booking[]>([])
   const [acceptedBookings, setAcceptedBookings] = useState<Booking[]>([])
+  const [cardsToShow, setCardsToShow] = useState<'pending' | 'accepted'>('pending')
 
   useEffect(() => {
     // const fetchBookings = async () => {
@@ -61,29 +46,33 @@ function SitterHomeScreen() {
     setAcceptedBookings(myAcceptedBookings)
   }, [])
 
-  const handleAcceptBooking = useCallback(async (bookingId: number) => {
-    const response = await axios.patch(
-      `${import.meta.env.VITE_CATCARE_SERVER_URL}/booking/answer-booking`,
-      {
-        bookingId: bookingId,
-        answerBooking: 'ACCEPTED'
-      },
-      {
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${getAuthTokenFromStorage()}`
+  const handleAcceptBooking = useCallback(
+    async (bookingId: number) => {
+      const response = await axios.patch(
+        `${import.meta.env.VITE_CATCARE_SERVER_URL}/booking/answer-booking`,
+        {
+          bookingId: bookingId,
+          answerBooking: 'ACCEPTED'
+        },
+        {
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${getAuthTokenFromStorage()}`
+          }
         }
+      )
+
+      if (response.status < 200 || response.status >= 300) {
+        toast.error('Não foi possível aceitar o booking.')
+        return
       }
-    )
 
-    if (response.status < 200 || response.status >= 300) {
-      toast.error('Não foi possível aceitar o booking.')
-      return
-    }
-
-    setPendingBookings((prev) => prev.filter((booking) => booking.id !== bookingId))
-    setAcceptedBookings((prev) => [...prev, response.data])
-  }, [])
+      const bookingToAccept = pendingBookings.find((booking) => booking.id === bookingId)
+      setAcceptedBookings((prev) => [...prev, bookingToAccept])
+      setPendingBookings((prev) => prev.filter((booking) => booking.id !== bookingId))
+    },
+    [pendingBookings, acceptedBookings]
+  )
 
   const handleRejectBooking = useCallback(async (bookingId: number) => {
     const response = await axios.patch(
@@ -110,58 +99,29 @@ function SitterHomeScreen() {
 
   return (
     <SitterHomeContainer>
-      <CardsList>
-        {pendingBookings.map((booking) => (
-          <BookingCard key={booking.id}>
-            <DateInfo>
-              {new Date(booking.startDate).toLocaleDateString()} - {new Date(booking.endDate).toLocaleDateString()}
-            </DateInfo>
-            <Header>
-              <Name>
-                <strong>Solicitante:</strong> {requestersData.find((user) => user.id === booking.requesterId)?.name}
-              </Name>
-            </Header>
-            <Address>
-              <strong>Endereço:</strong> {requestersData.find((user) => user.id === booking.requesterId)?.address}
-            </Address>
-            {[...booking.visits].map((visit, index) => (
-              <VisitWrapper key={visit.visitDate?.toISOString() ?? index}>
-                <VisitSummary expandIcon={<ArrowDropDown />}>{`Visita ${index + 1}`}</VisitSummary>
-                <VisitItem>
-                  <Info>
-                    <strong>Data:</strong> {visit.visitDate?.toLocaleDateString()}
-                  </Info>
-                  <Info>
-                    <strong>Observações:</strong>
-                    <Details>{visit.notes || <i>Sem observações.</i>}</Details>
-                  </Info>
-                </VisitItem>
-              </VisitWrapper>
-            ))}
-            <Notes>{booking.generalNotes}</Notes>
-            <Footer>
-              <Button
-                variant="filled"
-                color="#d32f2f"
-                size="sm"
-                fullWidth
-                onClick={() => handleRejectBooking(booking.id)}
-              >
-                Recusar
-              </Button>
-              <Button
-                variant="filled"
-                color="#00a128"
-                size="sm"
-                fullWidth
-                onClick={() => handleAcceptBooking(booking.id)}
-              >
-                Aceitar
-              </Button>
-            </Footer>
-          </BookingCard>
-        ))}
-      </CardsList>
+      <ButtonsWrapper>
+        <Button
+          onClick={() => setCardsToShow('pending')}
+          fullWidth
+          size="md"
+          variant={cardsToShow === 'pending' ? 'filled' : 'outline'}
+        >
+          Pendentes
+        </Button>
+        <Button
+          onClick={() => setCardsToShow('accepted')}
+          fullWidth
+          size="md"
+          variant={cardsToShow === 'accepted' ? 'filled' : 'outline'}
+        >
+          Aceitos
+        </Button>
+      </ButtonsWrapper>
+      <BookingsList
+        bookings={cardsToShow === 'pending' ? pendingBookings : acceptedBookings}
+        onAcceptBooking={handleAcceptBooking}
+        onRejectBooking={handleRejectBooking}
+      />
     </SitterHomeContainer>
   )
 }
